% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_edid_model.R
\name{estimate_edid_model}
\alias{estimate_edid_model}
\title{Estimate the efficient DiD model}
\usage{
estimate_edid_model(
  dat,
  ytilde,
  inf_func,
  cluster = NULL,
  biters = 1000,
  seed = 6
)
}
\arguments{
\item{dat}{The data frame returned by \code{\link[=prep_edid_data]{prep_edid_data()}}.}

\item{ytilde}{The list of Ytilde values returned by \code{\link[=get_ytilde]{get_ytilde()}}.}

\item{inf_func}{The list of influence function values returned by
\code{\link[=get_influence_func]{get_influence_func()}}.}

\item{cluster}{Optional boolean to override the default standard error
clustering behavior. The default \code{NULL} will cluster standard errors
if the data frame has a column named "\code{clust}", i.e. if a \code{cluster_var}
was specified in \code{\link[=prep_edid_data]{prep_edid_data()}}. This can be overridden by setting
this argument to \code{FALSE}, producing the default standard errors
(which are still clustered at the unit level). Setting the argument to
\code{TRUE} without having specified a cluster variable in \code{\link[=prep_edid_data]{prep_edid_data()}}
will throw an error.}

\item{biters}{An integer specifying number of bootstrap iterations to
perform.}

\item{seed}{An integer used to set the random seed for bootstrapping.}
}
\value{
A list with one element per treatment group and post-period (g, t). Each
element's name is a sequence of two time periods, the first denoting the
treatment group (g) and the second denoting the post-period (t). Note that
the treatment group numbers are equal to the treatment adoption times in the
original dataset minus anticipation periods. Each list element contains the
following sixteen items:
\itemize{
\item 1: The set of ATTs for the ATT(g,t), one for each combination of
not-yet-treated comparison groups (g') and their period 2's (t_pre).
\item 2: The efficiency weights for each ATT(g, t, g', t_pre) used to combine
the ATTs into the efficient ATT(g,t).
\item 3: The efficient influence function values for each unit.
\item 4: The efficient ATT(g,t).
\item 5-6: The bootstrapped and analytic standard errors respectively.
\item 7-8: The 95\% and 90\% simultaneous confidence band critical values
respectively.
\item 9-16: The 95\% and 90\% simultaneous confidence bands for bootstrapped and
analytic standard errors respectively.
}
}
\description{
Computes each ATT(g, t, g', t_pre), their efficiency weights, and combines
them into the final ATT(g,t)s as described in
\href{https://arxiv.org/abs/2506.17729}{Chen, Sant'Anna and Xie (2025)}.
Also computes standard errors, both analytic and bootstrapped, and 95\% and
90\% simultaneous confidence bands.
}
\details{
The bootstrapped standard errors and simultaneous confidence bands are
computed following the procedure described in
\href{https://www.sciencedirect.com/science/article/abs/pii/S0304407620303948}{Callaway and Sant'Anna (2021)}
with code adapted directly from the authors'
\href{https://bcallaway11.github.io/did/}{did package} that implements the
procedure outlined in the paper.
}
\examples{
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 3
)
ytilde_list <- get_ytilde(edid_data)
inf_func_list <- get_influence_func(edid_data, ytilde_list)

edid_mod <- estimate_edid_model(
  edid_data, ytilde_list, inf_func_list
)

# Adjust bootstrap iterations & random seed generation
edid_mod <- estimate_edid_model(
  edid_data,
  ytilde_list,
  inf_func_list,
  biters = 5000,
  seed = 66
)

# With clustered standard errors
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 3,
  cluster_var = "cluster"
)
ytilde_list <- get_ytilde(edid_data)
inf_func_list <- get_influence_func(edid_data, ytilde_list)

edid_mod <- estimate_edid_model(
  edid_data,
  ytilde_list,
  inf_func_list,
  cluster = TRUE
)
}
