% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edid_results_agg.R
\name{get_edid_results_agg}
\alias{get_edid_results_agg}
\title{Get aggregated efficient DiD event study or calendar time results}
\usage{
get_edid_results_agg(
  dat,
  mod,
  anticip,
  type = "es",
  res = NULL,
  cluster = FALSE,
  biters = 1000,
  seed = 6
)
}
\arguments{
\item{dat}{The data frame returned by \code{\link[=prep_edid_data]{prep_edid_data()}}.}

\item{mod}{The EDiD model returned by \code{\link[=estimate_edid_model]{estimate_edid_model()}}.}

\item{anticip}{The number of treatment anticipation periods. This must align
with the number specified in \code{\link[=prep_edid_data]{prep_edid_data()}}. The treatment group IDs
will be accordingly re-aligned with the original dataset and aggregated
treatment effects computed with \code{\link[=get_edid_results_agg]{get_edid_results_agg()}} will also be
computed accordingly. Unlike in \code{\link[=prep_edid_data]{prep_edid_data()}}, no default is
provided as a safety measure against the scenario where one forgets to
re-specify non-zero anticipation periods.}

\item{type}{A string indicating the type of aggregation, either \code{es} for
event study (default) or \code{cal} for calendar time.}

\item{res}{Optionally supply the results data frame returned by
\code{\link[=edid_get_results_attgt]{edid_get_results_attgt()}}. The default \code{NULL} will run
\code{\link[=edid_get_results_attgt]{edid_get_results_attgt()}} under the hood to obtain the data frame.}

\item{cluster}{Boolean specifying whether to cluster standard errors at a
level higher than the unit level. The default is set to \code{FALSE}, which
results in standard errors clustered at the unit level. A cluster variable
must have been specified in \code{\link[=prep_edid_data]{prep_edid_data()}} or the function will throw
an error.}

\item{biters}{An integer specifying number of bootstrap iterations to
perform.}

\item{seed}{An integer used to set the random seed for bootstrapping.}
}
\value{
A data frame with the number of rows equal to the number of time
periods for which treatment effects are aggregated. The data frame will
contain the following 14 columns:
\itemize{
\item \code{e} or \code{t}: The event (\code{e}) or calendar (\code{t}) time period.
\item \code{att}: The efficient ATT(g,e) or ATT(g,t) aggregated across all treatment
groups.
\item \code{se_boot} & \code{se_analytic}: The bootstrapped and analytic standard errors.
\item \code{crit_val_95} & \code{crit_val_90}: The 95\% and 90\% simultaneous confidence
band critical values obtained during the bootstrapping procedure.
\item \verb{ci_*_*_*}: The lower (upper) 95\% (90\%) bootstrapped (anlaytic)
simultaneous confidence bands.
}
}
\description{
Aggregates the efficient ATT(g,t)s into event study or calendar time results.
}
\examples{
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 3
)
ytilde_list <- get_ytilde(edid_data)
inf_func_list <- get_influence_func(edid_data, ytilde_list)
edid_mod <- estimate_edid_model(
  edid_data, ytilde_list, inf_func_list
)

# Obtain event study results directly after model estimation
res_es <- get_edid_results_agg(
  edid_data, edid_mod, anticip = 0
)

# Use individual ATT(g,t) results
res_attgt <- get_edid_results_attgt(
  edid_data, edid_mod, anticip = 0
)
res_es <- get_edid_results_agg(
  edid_data, edid_mod, anticip = 0, res = res_attgt
)

# Get calendar time aggregated results
res_cal <- get_edid_results_agg(
  edid_data, edid_mod, anticip = 0, type = 'cal'
)
}
