% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edid_results_attgt.R
\name{get_edid_results_attgt}
\alias{get_edid_results_attgt}
\title{Extract the efficient ATT(g,t) results from an EDiD model}
\usage{
get_edid_results_attgt(dat, mod, anticip)
}
\arguments{
\item{dat}{The data frame returned by \code{\link[=prep_edid_data]{prep_edid_data()}}.}

\item{mod}{The EDiD model returned by \code{\link[=estimate_edid_model]{estimate_edid_model()}}.}

\item{anticip}{The number of treatment anticipation periods. This must align
with the number specified in \code{\link[=prep_edid_data]{prep_edid_data()}}. The treatment group IDs
will be accordingly re-aligned with the original dataset and aggregated
treatment effects computed with \code{\link[=get_edid_results_agg]{get_edid_results_agg()}} will also be
computed accordingly. Unlike in \code{\link[=prep_edid_data]{prep_edid_data()}}, no default is
provided as a safety measure against the scenario where one forgets to
re-specify non-zero anticipation periods.}
}
\value{
A data frame with the number of rows equal to the number of efficient
ATT(g,t)s estimated. The data frame with contain the following 19
columns:
\itemize{
\item \code{g}: The treatment group, defined by the treatment adoption time in the
original dataset.
\item \code{t}: The time period.
\item \code{e}: The "event time," i.e. time relative to \code{g}.
\item \code{att}: The efficient ATT(g,t) estimate.
\item \code{se_boot} & \code{se_analytic}: The bootstrapped and analytic standard errors.
\item \code{crit_val_95} & \code{crit_val_90}: The 95\% and 90\% simultaneous confidence
band critical values obtained during the bootstrapping procedure.
\item \verb{ci_*_*_*}: The lower (upper) 95\% (90\%) bootstrapped (anlaytic)
simultaneous confidence bands.
\item \code{pi_g}: The proportion of the sample comprised of members of treatment
group \code{g}.
\item \code{q_ge}: The proportion of all units treated in event time \code{e} that belong
to treatment group \code{g}.
\item \code{q_gt}: The proportion of all units treated in time \code{t} that belong to
treatment group \code{g}. Units with anticipation periods in time \code{t} are
omitted from this calculation and have \code{NA} values for this variable.
}
}
\description{
Extracts the efficient ATT(g,t)s, standard errors, and simultaneous
confidence bands from an efficient DiD model and puts them in a data frame.
}
\examples{
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 3
)
ytilde_list <- get_ytilde(edid_data)
inf_func_list <- get_influence_func(edid_data, ytilde_list)
edid_mod <- estimate_edid_model(
  edid_data, ytilde_list, inf_func_list
)

res_attgt <- get_edid_results_attgt(
  edid_data, edid_mod, anticip = 0
)

# With an anticipation period
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 2,
  anticip = 1
)
ytilde_list <- get_ytilde(edid_data)
inf_func_list <- get_influence_func(edid_data, ytilde_list)
edid_mod <- estimate_edid_model(
  edid_data, ytilde_list, inf_func_list
)

res_attgt <- get_edid_results_attgt(
  edid_data, edid_mod, anticip = 1
)
}
