% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ytilde.R
\name{get_ytilde}
\alias{get_ytilde}
\title{Compute the Ytilde values for group, time and comparison}
\usage{
get_ytilde(dat)
}
\arguments{
\item{dat}{The data frame returned by \code{\link[=prep_data]{prep_data()}}.}
}
\value{
A list with one element per (g, t, g', t_pre). Each element contains a
vector the length of the number of units containing each unit's Ytilde
value. The name of each list element contains a sequence of four time
period numbers, separated by "\verb{_}". The numbers correspond to the
following:
\enumerate{
\item Treatment group (g). Note that these are equal to the treatment adoption
times in the original dataset minus anticipation periods.
\item Post-treatment period (t).
\item Not-yet-treated comparison group (g').
\item Period 2 for the not-yet-treated comparison group (t_pre in Chen et al.).
Note that this value is referred to as \code{tmid} in the code to avoid confusion
with the pre-treatment periods for the treatment group.
}
}
\description{
Computes the "Ytilde" function as described in
\href{https://arxiv.org/abs/2506.17729}{Chen, Sant'Anna and Xie (2025)} for
each combination of treatment group (g), post-treatment period (t) and
control group comparison (g', t_pre) set.
}
\examples{
edid_data <- prep_data(
  edid_ex_data,
  y = "outcome",
  id = "id",
  treat_time = "treat_adopt_time",
  time = "time",
  num_t_pre = 3
)

ytilde_list <- get_ytilde(edid_data)
}
