% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_edid.R
\name{plot_edid}
\alias{plot_edid}
\title{Plot EDiD results}
\usage{
plot_edid(
  res_dat,
  mod_type = NULL,
  ci = "95_boot",
  colors = "turbo",
  error_bar_width = 0.3,
  x_lim_padding = 0.2,
  text_size_base = NULL,
  text_size_title = NULL,
  text_size_legend = NULL,
  text_size_legend_title = NULL,
  text_font = NULL,
  gt_g = NULL,
  gt_ncol = NULL,
  gt_y_axes_same = TRUE
)
}
\arguments{
\item{res_dat}{A results data frame returned by \code{\link[=edid]{edid()}},
\code{\link[=get_edid_results_attgt]{get_edid_results_attgt()}}, or \code{\link[=get_edid_results_agg]{get_edid_results_agg()}}}

\item{mod_type}{Character string specifying the type of model you want to
plot. Options are \code{"attgt"}, \code{"es"}, and \code{"cal"}. If your results data
frame contains only one model type, you can leave the default \code{NULL}
option and the function will automatically plot that model type.}

\item{ci}{A character string specifying the confidence intervals you want to
display. The options are: \code{"95_boot"} (default), \code{"90_boot"},
\code{"95_analytic"}, \code{"90_analytic"}, \code{"boot"}, \code{"analytic"}, \code{"95"}, \code{"90"},
and \code{"all"}. \code{"boot"} and \code{"analytic"} will plot both the 95\% and 90\%
bootstrapped or analytic confidence bands respectively. \code{"95"} and \code{"90"}
will plot both the bootstrapped and analytic confidence bands at the 95\%
or 90\% confidence levels respectively. \code{"all"} will plot all four
confidence bands.}

\item{colors}{A character string or vector specifying the color palette.
There are three special string characters that will specify specific
palettes from the colorblind-friendly \code{viridis} color palettes: \code{"turbo"}
(default), \code{"viridis"}, and \code{"viridis_.75"} which is the viridis palette
with a truncated range that I find more visually appealing. You can also
specify your own custom palette by supplying a character vector the same
length as the number of requested confidence intervals. The order matters:
colors are ordered from largest confidence intervals to smallest. That
means the last item of the vector corresponds to the smallest confidence
interval, which will be in the foreground and therefore the dominant color
of the plot. The ATT points will also be this color.}

\item{error_bar_width}{A number specifying the width of the horizontal end
points of the error bars. If error bars are too wide, the first and/or
last time period's bars may extend beyond the x-axis limits and get
clipped, so you may need to adjust this setting accordingly.}

\item{x_lim_padding}{A number specifying the amount by which to extend the
x-axis limits in order to provide "padding" for the error bars to prevent
clipping. If your error bars are getting clipped (i.e., the first or last
time period's horizontal bar ends do not appear), you will need to
increase this value. You can also decrease the value if you have extra
unnecessary space at the ends of the x-axis.}

\item{text_size_base}{An integer specifying the ggplot2 base text size.
\code{NULL} defaults to \code{11}, the ggplot2 default. If the value is supplied
and the values to the other \verb{text_size_*} arguments are left \code{NULL}, the
text sizes of those elements will be updated to match the supplied base
text size.}

\item{text_size_title}{An integer specifying the plot title size. \code{NULL}
defaults to the same size as \code{text_size_base}.}

\item{text_size_legend}{An integer specifying the ggplot2 legend text size.
\code{NULL} defaults to two sizes smaller than \code{text_size_base}.}

\item{text_size_legend_title}{An integer specifying the ggplot2 legend title
size. \code{NULL} defaults to the same size as \code{text_size_base}.}

\item{text_font}{A string specifying the name of the font family to use.
\code{NULL} defaults to \code{sans}, the ggplot2 default.}

\item{gt_g}{This argument only applies to \code{mod_type = "attgt"}. A numeric
vector specifying a subset of treatment groups to plot. The default \code{NULL}
will plot all treatment groups.}

\item{gt_ncol}{This argument only applies to \code{mod_type = "attgt"}. An
integer specifying the number of columns in the plot grid, i.e. the final
output that is a grid of the individual ATT(g,t) plots. If the default
\code{NULL} is supplied, the function will auto-assign a number of cols as
follows: \code{1} if there are 1 or 2 treatment groups; \code{2} if there are 3 or 4
treatment groups; and \code{3} otherwise.}

\item{gt_y_axes_same}{This argument only applies to \code{mod_type = "attgt"}. A
boolean specifying whether y-axes should be consistent across all
individual ATT(g,t) plots. The default is \code{TRUE}.}
}
\value{
A ggplot2 plot. If \code{mod_type = "attgt"}, it returns a grid of
individual ATT(g,t) ggplots created by the patchwork package.
}
\description{
This function uses ggplot2 to plot the EDiD results. If you wish to plot
multiple individual treatment group results, the function uses the patchwork
package to fashion these individual plots into a grid. You can choose to
plot any or all of the confidence intervals. Any anticipation years will be
plotted in black and post-treatment years will be plotted in color.
}
\examples{
edid_res <- edid(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 3
)

# Plot individual ATT(g,t)'s with the 95\% bootstrapped CI
plot_edid(edid_res, mod_type = "attgt")

# Plot a subset of individual ATT(g,t)'s with all CIs
plot_edid(
  edid_res,
  mod_type = "attgt",
  ci = "all",
  gt_g = c(5, 6, 7)
)

# Plot event study with 95\% and 90\% bootstrapped CIs
# and the viridis color scheme
plot_edid(
  edid_res,
  mod_type = "es",
  ci = "boot",
  colors = "viridis"
)

# Plot calendar time with bootstrapped and analytic 95\% CIs,
# a custom color palette, and smaller plot text
plot_edid(
  edid_res,
  mod_type = "cal",
  ci = "95",
  colors = c("deeppink2", "deepskyblue2"),
  text_size_base = 10
)

# Plot an event study with an anticipation period
edid_res_anticip <- edid(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 2,
  anticip = 1
)

plot_edid(edid_res_anticip, mod_type = "es")

# Plot an event study with a results data frame
# containing only event study results
edid_res_es <- edid(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 2,
  anticip = 1,
  get_attgt = FALSE,
  get_cal = FALSE
)

plot_edid(edid_res_es)
}
