% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_edid_data.R
\name{prep_edid_data}
\alias{prep_edid_data}
\title{Prepare a dataset for EDiD modeling}
\usage{
prep_edid_data(
  dat,
  y_var,
  treat_time_var,
  id_var,
  time_var,
  num_t_pre,
  cluster_var = NULL,
  anticip = 0
)
}
\arguments{
\item{dat}{A data frame with the data you want to model.}

\item{y_var}{A character string with the name of the outcome variable name
in the data frame.}

\item{treat_time_var}{A string with the name of a numeric variable in the
data frame that indicates the first period of treatment. Never treated
units should have a value of 0 on this variable.}

\item{id_var}{A string with the unit-level ID variable name in the
data frame. The variable can be numeric or character.}

\item{time_var}{A string with name of the numeric time variable in the
data frame.}

\item{num_t_pre}{An integer specifying the number of pre-treatment
periods to use for computing treatment effects.}

\item{cluster_var}{An optional string with the name of a variable in the
data frame to cluster standard errors on. \code{NULL} defaults to clustering at
the unit level. The variable can be numeric or character.}

\item{anticip}{An integer specifying the number of pre-treatment
anticipation periods. Treatment adoption will be re-defined as the first
anticipation period for modeling purposes. The treatment adoption period
will be re-aligned with the initial dataset when extracting results and
computing aggregated effects. The default is set to \code{0}.}
}
\value{
A wide-format, balanced-panel data frame that can be used with the
efficient DiD estimator functions. The row length will be the number of
units with data observed in all time periods. The columns will include:
\itemize{
\item \code{id}: each unit's ID
\item \code{g}: the unit's treatment group
\item \code{y[time]}: one column per time period containing outcome data for that
period
\item \verb{g[treatment adoption time]}: one column per treatment group containing a
binary indicator of treatment group membership.
}
}
\description{
This takes a panel dataset with the data you want to model and returns a
wide dataset with the variables needed for modeling, renamed to work with
the modeling functions.
}
\examples{
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  treat_time_var = "treat_adopt_time",
  id_var = "id",
  time_var = "time",
  num_t_pre = 3
)

# Clustered standard errors
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  treat_time_var = "treat_adopt_time",
  id_var = "id",
  time_var = "time",
  num_t_pre = 3,
  cluster_var = "cluster"
)

# Specify anticipation
edid_data <- prep_edid_data(
  edid_ex_data,
  y_var = "outcome",
  treat_time_var = "treat_adopt_time",
  id_var = "id",
  time_var = "time",
  num_t_pre = 2,
  anticip = 1
)
}
