% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_edid.R
\name{plot_edid}
\alias{plot_edid}
\title{Plot EDiD results}
\usage{
plot_edid(
  res_dat,
  type = NULL,
  ci = "95_boot",
  colors = "turbo",
  title = NULL,
  ylab = "ATT",
  xlab = NULL,
  point_size = NULL,
  error_bar_vline_width = NULL,
  error_bar_hline_width = 0.3,
  xlim_padding = 0.2,
  text_size_base = NULL,
  text_size_title = NULL,
  text_size_legend_title = NULL,
  text_size_legend_text = NULL,
  show_legend = TRUE,
  text_font = NULL,
  anticip_color = NULL,
  x_breaks = NULL,
  gt_g = NULL,
  gt_ci_order = NULL,
  gt_ncol = NULL,
  gt_y_axes_same = TRUE
)
}
\arguments{
\item{res_dat}{A results data frame returned by \code{\link[=edid]{edid()}},
\code{\link[=get_results_attgt]{get_results_attgt()}}, or \code{\link[=get_results_agg]{get_results_agg()}}}

\item{type}{Character string specifying the type of model you want to
plot. Options are \code{"attgt"}, \code{"es"}, and \code{"cal"}. If your results data
frame contains only one model type, you can leave the default \code{NULL}
option and the function will automatically plot that model type.}

\item{ci}{A character string specifying the confidence intervals you want to
display. The options are: \code{"95_boot"} (default), \code{"90_boot"},
\code{"95_analytic"}, \code{"90_analytic"}, \code{"boot"}, \code{"analytic"}, \code{"95"}, \code{"90"},
and \code{"all"}. \code{"boot"} and \code{"analytic"} will plot both the 95\% and 90\%
bootstrapped or analytic confidence bands respectively. \code{"95"} and \code{"90"}
will plot both the bootstrapped and analytic confidence bands at the 95\%
or 90\% confidence levels respectively. \code{"all"} will plot all four
confidence bands.}

\item{colors}{A character string or vector specifying the color palette.
There are three special string characters that will specify specific
palettes from the colorblind-friendly \code{viridis} color palettes: \code{"turbo"}
(default), \code{"viridis"}, and \code{"viridis_.75"} which is the viridis palette
with a truncated range that I find more visually appealing. You can also
specify your own custom palette by supplying a character vector the same
length as the number of requested confidence intervals. The order matters:
colors are ordered from largest confidence intervals to smallest. That
means the last item of the vector corresponds to the smallest confidence
interval, which will be in the foreground and therefore the dominant color
of the plot. The ATT points will also be this color.}

\item{title}{A string that will become the plot title. The default \code{NULL}
omits a title.}

\item{ylab}{A string to be used as the y-axis lable. The default is \code{"ATT"}.}

\item{xlab}{A string to be used as the x-axis label. \code{NULL} defaults to
\code{"Time"} for \code{"attgt"} and \code{"cal"} plots and \code{"Time Since Event"} for
\code{"es"} plots.}

\item{point_size}{A number specifying the size the points that represent
the ATTs, generated by \code{ggplot2::geom_point()}. \code{NULL} defaults to \code{1.5},
the ggplot2 default.}

\item{error_bar_vline_width}{A number specifying the width of the vertical
error bar lines. \code{NULL} defaults to \code{0.5}, the ggplot2 default.}

\item{error_bar_hline_width}{A number specifying the width of the horizontal
end points of the error bars. If error bars are too wide, the first and/or
last time period's bars may extend beyond the x-axis limits and get
clipped, so you may need to adjust this setting accordingly.}

\item{xlim_padding}{A number specifying the amount by which to extend the
x-axis limits in order to provide "padding" for the error bars to prevent
clipping. If your error bars are getting clipped (i.e., the first or last
time period's horizontal bar ends do not appear), you will need to
increase this value. You can also decrease the value if you have extra
unnecessary space at the ends of the x-axis.}

\item{text_size_base}{A number specifying the ggplot2 base text size.
\code{NULL} defaults to \code{11}, the ggplot2 default. If the value is supplied
and the values to the other \verb{text_size_*} arguments are left \code{NULL}, the
text sizes of those elements will be updated to match the supplied base
text size.}

\item{text_size_title}{A number specifying the plot title size. \code{NULL}
defaults to\code{text_size_base + 1}, the ggplot2 default, for \code{type = "es"}
and \code{type = "cal"}. For \code{type = "aggt"}, this parameter corresponds to the
size of the individual ATT(g,t) plot titles rather than the overall plot
grid title. In this case, \code{NULL} defaults to the same size as
\code{text_size_base}. The plot grid title will always be set to
\code{text_size_base + 1} (the ggplot2 default).}

\item{text_size_legend_title}{A number specifying the ggplot2 legend title
size. \code{NULL} defaults to the same size as \code{text_size_base}.}

\item{text_size_legend_text}{A number specifying the ggplot2 legend text size.
\code{NULL} defaults to two sizes smaller than \code{text_size_base}.}

\item{show_legend}{A boolean specifying whether to display the confidence
interval legend. Default is \code{TRUE}.}

\item{text_font}{A string specifying the name of the font family to use.
\code{NULL} defaults to \code{sans}, the ggplot2 default.}

\item{anticip_color}{A string specifying the color of anticipation periods.
\code{NULL} defaults to \code{"black"}.}

\item{x_breaks}{A numeric vector specifying custom x-axis breaks. The
default \code{NULL} will use the default ggplot2 algorithm.}

\item{gt_g}{This argument only applies to \code{type = "attgt"}. A numeric
vector specifying a subset of treatment groups to plot. The default \code{NULL}
will plot all treatment groups.}

\item{gt_ci_order}{This argument only applies to \code{type = "attgt"}. A
character vector specifying a custom order for the confidence intervals
to be entered into the plot. The vector must contain the names of all
of the confidence intervals you are plotting and nothing else. The names
are \code{"95_boot"}, \code{"90_boot"}, \code{"95_anlaytic"}, and \code{"90_analytic"}. This
is only necessary if you want a different order than the default, which is
\code{c("95_boot", "95_analytic", "90_boot", "90_analytic")}. Unlike the event
study and calendar time plots, an ordering algorithm cannot be used to
find the order of confidence interval size because it may differ across
treatment groups, and the order must be the same across all treatment
groups. Therefore the default makes an assumption about the order, and
that order can only be changed with this argument.}

\item{gt_ncol}{This argument only applies to \code{type = "attgt"}. An
integer specifying the number of columns in the plot grid, i.e. the final
output that is a grid of the individual ATT(g,t) plots. If the default
\code{NULL} is supplied, the function will auto-assign a number of cols as
follows: \code{1} if there are 1 or 2 treatment groups; \code{2} if there are 3 or 4
treatment groups; and \code{3} otherwise.}

\item{gt_y_axes_same}{This argument only applies to \code{type = "attgt"}. A
boolean specifying whether y-axes should be consistent across all
individual ATT(g,t) plots. The default is \code{TRUE}.}
}
\value{
A ggplot2 plot. If \code{type = "attgt"}, it returns a grid of
individual ATT(g,t) ggplots created by the patchwork package.
}
\description{
This function uses ggplot2 to plot the EDiD results. If you want to plot
multiple individual treatment group results, the function uses the patchwork
package to fashion these individual plots into a grid. You can choose to
plot any or all of the confidence intervals.
}
\details{
Any anticipation years will be plotted in black by default or a custom
user-specified color and post-treatment years will be plotted in color.
An algorithm finds the size order of the confidence intervals and enters
them into the plot starting with the largest and ending with the smallest,
so that the confidence interval in the foreground of the plot is the
smallest. The color of point representing the ATT is colored the same as
this foreground confidence interval. The algorithm finds the relative size
of the confidence intervals at each time period and assigns order based on
the proportion of time periods in which each confidence interval is larger
than another. For example, if the 95\% analytic confidence interval is larger
than the 90\% bootstrapped confidence interval in 70\% of the time periods, it
will get the higher order assignment.
}
\examples{
edid_res <- edid(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 3
)

# Plot individual ATT(g,t)'s with the 95\% bootstrapped CI
plot_edid(edid_res, type = "attgt")

# Plot a subset of individual ATT(g,t)'s with all CIs
plot_edid(
  edid_res,
  type = "attgt",
  ci = "all",
  gt_g = c(5, 6, 7)
)

# Plot event study with 95\% and 90\% bootstrapped CIs
# and the viridis color scheme
plot_edid(
  edid_res,
  type = "es",
  ci = "boot",
  colors = "viridis"
)

# Plot calendar time with bootstrapped and analytic 95\% CIs,
# a custom color palette, and smaller plot text
plot_edid(
  edid_res,
  type = "cal",
  ci = "95",
  colors = c("deeppink2", "deepskyblue2"),
  text_size_base = 10
)

# Plot an event study with an anticipation period
edid_res_anticip <- edid(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 2,
  anticip = 1
)

plot_edid(edid_res_anticip, type = "es")

# Plot an event study with a results data frame
# containing only event study results
edid_res_es <- edid(
  edid_ex_data,
  y_var = "outcome",
  id_var = "id",
  treat_time_var = "treat_adopt_time",
  time_var = "time",
  num_t_pre = 2,
  anticip = 1,
  get_attgt = FALSE,
  get_cal = FALSE
)

plot_edid(edid_res_es)
}
